Count words of a sentence
A sendtence is made up of a group of words. Each word is a sequence of letters, ('a'-'z', 'A'-'Z'), that may contain one or more hyphens and may end in a punctuation mark: period(.), comma(,), question mark(?), or exclamenation point(!). Words will be separated by one or more white space characters.  Hyphens join two words into one and should be retained while the other punctuation marks should be stripped.  Determine the number of words in a given sentence